==> 01-setup-clip.ipynb <==
# Unsplash Image Search

The project allows you to search images on Unsplash by using a natural words description. It is powered by OpenAI's [CLIP model](https://github.com/openai/CLIP).

Use this notebook to setup your environment.

## Install the Dependencies

To install the project's dependencies (including those for CLIP) run the following in the terminal. It is a good idea to create a virtual environment.

```
pip install -r requirements.txt
```

## Clone the CLIP repository and copy the code

# Clone the CLIP repository
!git clone https://github.com/openai/CLIP.git

# Move the CLIP source files and the vocabulary in the root directory. 
# Unfortunately, the CLIP code is not organized as a module, so it cannot be imported easily
!mv CLIP/*.py .
!mv CLIP/*.gz .
